// dashboard/src/api/models/address.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */
import type { AddressCountry } from "./addressCountry";

export interface Address {
  /**
   * @maxLength 255
   * @nullable
   */
  line1?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  line2?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  locality?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  state?: string | null;
  /**
   * @maxLength 255
   * @nullable
   */
  postal_code?: string | null;
  /** @nullable */
  country?: AddressCountry;
}